pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = ""
authors = ["smpfy <arsmpfy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fire = "^0.5.0"
numpy = "^1.26.2"
torch = "^2.1.1"
torchvision = "^0.16.1"
onnx = "^1.15.0"
onnxscript = "^0.1.0.dev20231213"
onnxruntime = "^1.16.3"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
dvc = "^3.33.4"
dvc-gdrive = "^2.20.0"
pre-commit = "^3.6.0"
black = "^23.12.0"
isort = "^5.13.2"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
